server/requirements.txt
greenlet==1.1.2
paho-mqtt==1.6.1
psycopg2-binary==2.9.3
pydantic==1.9.1
SQLAlchemy==1.4.39
typing-extensions==4.3.0
